requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
db-dtypes==1.2.0
defusedxml==0.7.1
filelock==3.14.0
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
multidict==6.0.5
numpy==2.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas-gbq==0.23.1
parsel==1.9.1
Protego==0.3.1
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydata-google-auth==1.8.2
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.21.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==5.1.2
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
w3lib==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.4.post2
